feat: modifi test_deposit_into_strategy_with_signature
